Add "open source" and disclaimer in license notice
Make the following changes in the license notice that appears
when the script is run with the `-v` or `--version` option.

  1. Use the term "free and open source software" instead of
     "free software". Make this change in the README as well.
  2. Add disclaimer of warranty.
  3. Wrap the license notice at 65 characters, so that the
     license text is almost justified.

Additionally, define script level variables named AUTHOR,
LICENSE_URL and SUPPORT_URL and use them in the help and version
messages, so that these values are easy to modify later, should
any of these details change.