Fix use of random number to name STDIN file
When the hyphen, i.e. '-', argument is used to read the standard input
and edit it, a temporary STDIN file is created to save the text read
from the standard input. The path of this file is supposed to follow the
format /tmp/vimer/_STDIN_RANDOM_YYYY-MM-DD_HH-MM-SS where RANDOM is a
random number, YYYY-MM-DD is the current date and HH-MM-SS is the
current time.

However, due to a bug, RANDOM was missing from the filename. Fix this
bug in this commit.