Use sequence number in STDIN filename
When the hyphen, i.e. '-', argument is used to read the standard input
and edit it, a temporary STDIN file is created to save the text read
from the standard input. To ensure that two vimer instances can edit
text from the standard input simultaneously, the STDIN filenames need to
be unique. To ensure this, the filename of the STDIN file was chosen to
be of _STDIN_RANDOM_YYYY-MM-DD_HH-MM-SS format where RANDOM is a random
number.

The random number is supposed to ensure that two STDIN files created
within the same second have different filenames. However, this does not
work as intended because the current time (with one-second resolution)
is used to set the seed of the random number generator. As a result,
the random number used in the filename does not vary within a second.
Therefore, two STDIN files created within the same second have the same
filenames. The script goes into a loop to generate more filenames until
one is found that is unique. This loop succeeds in generating a new
unique filename only at the next second. This leads to a maximum of
one-second delay before the standard input can be edited when another
Vimer instance has already begun editing the standard input in the same
second.

Fix this issue by changing the format of the filename of STDIN file to
_STDIN_YYYY-MM-DD_HH-MM-SS_SEQ where SEQ is a sequence number. Start
with a sequence number of 0. If an STDIN filename with that name already
exists, just increment the sequence number by 1 and try again, and so on
until a unique STDIN filename is found. Therefore the filename
generation loop no longer has to wait for a second.